pyproject.toml
[project]
name = "qwenections"
version = "0.1.0"
description = "Fine-tune Qwen model to play NYT Connections"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "datasets>=2.14.0",
    "peft>=0.8.0",
    "accelerate>=0.27.0",
    "bitsandbytes>=0.41.0",
    "trl>=0.7.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
    "tqdm>=4.66.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

